// frontend-admin/src/api/generated/model/missingMonth.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API â€” RKSV-compliant POS / Admin / shared Auth surfaces
 * OpenAPI spec version: v1
 */

export interface MissingMonth {
  deadline: string;
  isOverdue: boolean;
  month: number;
  year: number;
}