// frontend-admin/src/api/generated/model/monatsbelegRegisterStatusItemDto.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API - RKSV uyumlu kasa sistemi
 * OpenAPI spec version: v1
 */
import type { MonatsbelegStatusDto } from './monatsbelegStatusDto';

export interface MonatsbelegRegisterStatusItemDto {
  cashRegisterId?: string;
  status?: MonatsbelegStatusDto;
}